Version A1.0
Completed, proof-of-concept alpha version 1.0.  Currently, the arduino
beacon is working, the beacon scanner in the app is working, detected
beacons enter a win or a loss into the SQLite database, and that data is
then reflected on the graph.

Currently known bugs:
-Not all phones communicate with the arduino correctly, but the arduino
will work with some phones.  Cause unknown at this time.  Arduino can be
simulated with a beacon emulator.
-App must be launched once to gain permissions, then closed and
relaunched again before beacons will be detected
-Graph will not display most recent entry, but will display all entries
prior.

-I can say with 100% confidence that there will be other bugs we have
not noticed yet, some of them will probably be really bad.  This is, at
this time, a proof-of-concept prototype release.  But it's a release
nonetheless!